Django_project2.0/pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = "django_project"
authors = ["Olga Tsuban <tsubanolga@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pymongo = "^4.6.1"
scrapy = "^2.11.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
